+++ src/com/kennesaw/cpumodule/DmaChannel.java
@@ -3,22 +3,22 @@ package com.kennesaw.cpumodule;
 import java.util.Arrays;
 
 public class DmaChannel {
-    private int pseudoRAM[];
+    private long pseudoRAM[];
 
     public DmaChannel() {
-        pseudoRAM = new int[1024];
+        pseudoRAM = new long[1024];
     }
 
     private int effectiveAddr(int addr) {
         return addr / 4;
     }
 
-    public int readRAM(int addr) {
+    public long readRAM(int addr) {
         addr = effectiveAddr(addr);
         return pseudoRAM[addr];
     }
 
-    public void writeRAM(int addr, int val) {
+    public void writeRAM(int addr, long val) {
         addr = effectiveAddr(addr);
         pseudoRAM[addr] = val;
     }
